rename bulk result upload related files
