+++ backend/src/test/java/gov/cdc/usds/simplereport/integration/UploadTestResultsIntegrationTest.java
@@ -17,11 +17,12 @@ import com.github.tomakehurst.wiremock.client.WireMock;
 import com.okta.commons.http.MediaType;
 import gov.cdc.usds.simplereport.api.BaseAuthenticatedFullStackTest;
 import gov.cdc.usds.simplereport.test_util.SliceTestConfiguration;
-import gov.cdc.usds.simplereport.utils.BulkUploadResultsToFhir;
-import gov.cdc.usds.simplereport.utils.BulkUploadResultsToFhirTest;
+import gov.cdc.usds.simplereport.utils.TestResultBulkUploadToFhir;
+import gov.cdc.usds.simplereport.utils.TestResultBulkUploadToFhirTest;
 import gov.cdc.usds.simplereport.utils.DateGenerator;
 import gov.cdc.usds.simplereport.utils.TokenAuthentication;
 import gov.cdc.usds.simplereport.utils.UUIDGenerator;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
@@ -30,6 +31,7 @@ import java.util.Date;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.UUID;
+
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -50,94 +52,99 @@ import org.springframework.test.web.servlet.MockMvc;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @TestPropertySource(properties = {"data-hub.fhir-enabled=true"})
 class UploadTestResultsIntegrationTest extends BaseAuthenticatedFullStackTest {
-  @Autowired private MockMvc mockMvc;
+    @Autowired
+    private MockMvc mockMvc;
 
-  @MockBean private TokenAuthentication _tokenAuth;
+    @MockBean
+    private TokenAuthentication _tokenAuth;
 
-  @MockBean private DateGenerator dateGenerator;
+    @MockBean
+    private DateGenerator dateGenerator;
 
-  @MockBean private UUIDGenerator uuidGenerator;
+    @MockBean
+    private UUIDGenerator uuidGenerator;
 
-  @SpyBean private BulkUploadResultsToFhir bulkUploadResultsToFhir;
+    @SpyBean
+    private TestResultBulkUploadToFhir testResultBulkUploadToFhir;
 
-  @BeforeAll
-  void setup() throws IOException {
-    Date date = Date.from(Instant.parse("2023-05-24T19:33:06.472Z"));
-    when(dateGenerator.newDate()).thenReturn(date);
-    when(uuidGenerator.randomUUID())
-        .thenReturn(UUID.fromString("5c5cc8fd-7001-4ac2-9340-541d065eca87"));
+    @BeforeAll
+    void setup() throws IOException {
+        Date date = Date.from(Instant.parse("2023-05-24T19:33:06.472Z"));
+        when(dateGenerator.newDate()).thenReturn(date);
+        when(uuidGenerator.randomUUID())
+                .thenReturn(UUID.fromString("5c5cc8fd-7001-4ac2-9340-541d065eca87"));
 
-    var properties = new Properties();
-    // short commit id
-    properties.setProperty("commit.id.abbrev", "CommitID");
-    properties.setProperty("commit.time", "1688565766");
-    ReflectionTestUtils.setField(
-        bulkUploadResultsToFhir, "gitProperties", new GitProperties(properties));
+        var properties = new Properties();
+        // short commit id
+        properties.setProperty("commit.id.abbrev", "CommitID");
+        properties.setProperty("commit.time", "1688565766");
+        ReflectionTestUtils.setField(
+                testResultBulkUploadToFhir, "gitProperties", new GitProperties(properties));
 
-    var responseFile =
-        getClass().getClassLoader().getResourceAsStream("responses/datahub-response.json");
+        var responseFile =
+                getClass().getClassLoader().getResourceAsStream("responses/datahub-response.json");
 
-    var mockResponse = IOUtils.toString(responseFile, StandardCharsets.UTF_8);
+        var mockResponse = IOUtils.toString(responseFile, StandardCharsets.UTF_8);
 
-    // reports covid
-    stubFor(
-        WireMock.post(urlEqualTo("/api/reports?processing=async"))
-            .willReturn(
-                WireMock.aResponse()
-                    .withStatus(HttpStatus.OK.value())
-                    .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                    .withBody(mockResponse)));
+        // reports covid
+        stubFor(
+                WireMock.post(urlEqualTo("/api/reports?processing=async"))
+                        .willReturn(
+                                WireMock.aResponse()
+                                        .withStatus(HttpStatus.OK.value())
+                                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                                        .withBody(mockResponse)));
 
-    // gets auth token
-    stubFor(
-        WireMock.post(urlEqualTo("/api/token"))
-            .withRequestBody(
-                equalTo(
-                    "scope=simple_report.*.report&grant_type=client_credentials&client_assertion_type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer&client_assertion=null"))
-            .willReturn(
-                WireMock.aResponse()
-                    .withStatus(HttpStatus.OK.value())
-                    .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                    .withBody("{\"access_token\": \"12345\"}")));
+        // gets auth token
+        stubFor(
+                WireMock.post(urlEqualTo("/api/token"))
+                        .withRequestBody(
+                                equalTo(
+                                        "scope=simple_report.*.report&grant_type=client_credentials&client_assertion_type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer&client_assertion=null"))
+                        .willReturn(
+                                WireMock.aResponse()
+                                        .withStatus(HttpStatus.OK.value())
+                                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                                        .withBody("{\"access_token\": \"12345\"}")));
 
-    // submits the FHIR bundles
-    stubFor(
-        WireMock.post(urlEqualTo("/api/waters"))
-            .willReturn(
-                WireMock.aResponse()
-                    .withStatus(HttpStatus.OK.value())
-                    .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                    .withBody(mockResponse)));
-  }
+        // submits the FHIR bundles
+        stubFor(
+                WireMock.post(urlEqualTo("/api/waters"))
+                        .willReturn(
+                                WireMock.aResponse()
+                                        .withStatus(HttpStatus.OK.value())
+                                        .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                                        .withBody(mockResponse)));
+    }
 
-  @Test
-  void CSVUpload() throws Exception {
+    @Test
+    void CSVUpload() throws Exception {
 
-    var sampleFhirMessage =
-        IOUtils.toString(
-            Objects.requireNonNull(
-                getClass()
-                    .getClassLoader()
-                    .getResourceAsStream("fhir/bundles-upload-integration-testing.ndjson")),
-            StandardCharsets.UTF_8);
+        var sampleFhirMessage =
+                IOUtils.toString(
+                        Objects.requireNonNull(
+                                getClass()
+                                        .getClassLoader()
+                                        .getResourceAsStream("fhir/bundles-upload-integration-testing.ndjson")),
+                        StandardCharsets.UTF_8);
 
-    InputStream input = loadCsv("testResultUpload/test-results-upload-integration.csv");
-    var file =
-        new MockMultipartFile(
-            "file",
-            "test-results-upload-integration.csv",
-            TEXT_CSV_CONTENT_TYPE,
-            input.readAllBytes());
+        InputStream input = loadCsv("testResultUpload/test-results-upload-integration.csv");
+        var file =
+                new MockMultipartFile(
+                        "file",
+                        "test-results-upload-integration.csv",
+                        TEXT_CSV_CONTENT_TYPE,
+                        input.readAllBytes());
 
-    mockMvc.perform(multipart(RESULT_UPLOAD).file(file)).andExpect(status().isOk());
+        mockMvc.perform(multipart(RESULT_UPLOAD).file(file)).andExpect(status().isOk());
 
-    verify(
-        exactly(1),
-        postRequestedFor(urlEqualTo("/api/waters"))
-            .withRequestBody(equalToJson(sampleFhirMessage, false, false)));
-  }
+        verify(
+                exactly(1),
+                postRequestedFor(urlEqualTo("/api/waters"))
+                        .withRequestBody(equalToJson(sampleFhirMessage, false, false)));
+    }
 
-  private InputStream loadCsv(String csvFile) {
-    return BulkUploadResultsToFhirTest.class.getClassLoader().getResourceAsStream(csvFile);
-  }
+    private InputStream loadCsv(String csvFile) {
+        return TestResultBulkUploadToFhirTest.class.getClassLoader().getResourceAsStream(csvFile);
+    }
 }
